[ doc ] PagingStateHelper, PagingStateManagerImpl
Author: @Kenny50
