--- app/src/main/java/com/kenny/githubreposearch/util/paging/PagingStateManager.kt
@@ -1,80 +1,0 @@
-package com.kenny.githubreposearch.util.paging
-
-import androidx.paging.PagingData
-import androidx.paging.cachedIn
-import com.kenny.githubreposearch.data.remote.paging_source.UiAction
-import com.kenny.githubreposearch.data.remote.paging_source.UiState
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.*
-import kotlinx.coroutines.launch
-
-interface PagingStateManager {
-}
-
-class PagingStateManagerImpl<T : Any>(
-    initialQuery: String,
-    lastQueryScrolled: String,
-    viewModelScope: CoroutineScope,
-    pagingSource: (String) -> Flow<PagingData<T>>
-) : PagingStateManager {
-    /**
-     * Stream of immutable states representative of the UI.
-     */
-    val state: StateFlow<UiState>
-
-    val pagingDataFlow: Flow<PagingData<T>>
-
-    /**
-     * Processor of side effects from the UI which in turn feedback into [state]
-     */
-    val accept: (UiAction) -> Unit
-
-    init {
-        val actionStateFlow = MutableSharedFlow<UiAction>()
-
-        val searches = actionStateFlow
-            .filterIsInstance<UiAction.Search>()
-            .distinctUntilChanged()
-            .onStart { emit(UiAction.Search(query = initialQuery)) }
-
-        val queriesScrolled = actionStateFlow
-            .filterIsInstance<UiAction.Scroll>()
-            .distinctUntilChanged()
-            // This is shared to keep the flow "hot" while caching the last query scrolled,
-            // otherwise each flatMapLatest invocation would lose the last query scrolled,
-            .shareIn(
-                scope = viewModelScope,
-                started = SharingStarted.WhileSubscribed(stopTimeoutMillis = 5000),
-                replay = 1
-            )
-            .onStart { emit(UiAction.Scroll(currentQuery = lastQueryScrolled)) }
-
-        pagingDataFlow = searches
-            .filterNot { it.query.isBlank() }
-            .flatMapLatest { pagingSource(it.query) }
-            .cachedIn(viewModelScope)
-
-        //ui action的兩個flow會造成  ui state的改變，當兩個query一樣時，表示用戶滾動中
-        state = combine(
-            searches,
-            queriesScrolled,
-            ::Pair
-        ).map { (search, scroll) ->
-            UiState(
-                query = search.query,
-                lastQueryScrolled = scroll.currentQuery,
-                // If the search query matches the scroll query, the user has scrolled
-                hasNotScrolledForCurrentSearch = search.query != scroll.currentQuery
-            )
-        }
-            .stateIn(
-                scope = viewModelScope,
-                started = SharingStarted.WhileSubscribed(stopTimeoutMillis = 5000),
-                initialValue = UiState()
-            )
-
-        accept = { action ->
-            viewModelScope.launch { actionStateFlow.emit(action) }
-        }
-    }
-}
